Add in support for a collector atom
We have to use fixed-size arrays, which means our connectivity
matrix will include a lot of bonds that the are not real
(i.e., only there to keep the array a fixed size).
This commit adds a fake node to the observed state
that is the source and sink for all of these fake bonds.
This fake node provides another benefit: ensures that
the maximum value of the connectivity is always the same,
which means tf.gather will all have the same shape.